overlay: init: urfkill: override to initialize wlan driver
Apparantly urfkill has a pre-start which ensures that services
inside the android container are fully started.

Our wlan driver needs fs access to initialize and load the
firmware files.
To signal fs access, the wifi HAL in android writes to /dev/wlan.
However this does not happen with halium and we need to manually
write to /dev/wlan.

Change-Id: I539bbc5261f001b77ef904691c0b6005507cb187
Signed-off-by: Alexander Martinz <alex@amartinz.at>